requirements.txt
Pillow
pyCLI